Rename fplApi.ts to api.ts
