+++ src/index.ts
@@ -1,5 +1,5 @@
+export * from './api';
 export * from './errors';
-export * from './fplAPI';
 export * from './FHIRDefinitions';
 export * from './load';
 export * from './utils';
